Use google-compute-engine package instead of google-daemon and google-startup-scripts packages, as per https://bugs.freebsd.org/bugzilla/show_bug.cgi?id=221714
Update commands echoed at the bottom to work with SDK version 169.

Use built-in growfs instead of adding yet another package.

Disable TSC timecounter so it automatically picks the best ACPI
timer. In my tests it is always ACPI-fast, but the original was trying
to use ACPI-safe, so there might still be some instances that get that
timer.

I always run with the serial port connected in the VM configuration,
so having an autoboot_delay is sensible. Also, I like the beastie logo
on the serial port. Works great!

Fix the ntp.conf so that the metadata server can set the time.

When disabling the virtual consoles, leave the serial port
'onifconsole' alone.