remove sdk package; fix error in echo
The cloud SDK package is not strictly necessary. There is only one
references to gsutil in the compute-engine software and it appears
optional. Not 100% sure what will appen when a start script is
specified for the VM, but I don't use that.

Also, need to quote the description in the final command that is
echoed.